+++ python/cuxfilter/charts/datashader/__init__.py
@@ -1,1 +1,8 @@
-from .datashader import scatter_geo, scatter, line, heatmap, stacked_lines
+from .datashader import (
+    scatter_geo,
+    scatter,
+    line,
+    heatmap,
+    stacked_lines,
+    graph,
+)